# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeLog": {
        "description": "Model for change logs",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "match_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Confidence"
          },
          "match_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Strategy"
          },
          "need_human_review": {
            "default": false,
            "title": "Need Human Review",
            "type": "boolean"
          },
          "prod_node": {
            "$ref": "#/components/schemas/Node"
          },
          "prop_changes": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {},
                {}
              ],
              "type": "array"
            },
            "title": "Prop Changes",
            "type": "object"
          },
          "staging_node": {
            "$ref": "#/components/schemas/Node"
          }
        },
        "required": [
          "prop_changes",
          "staging_node",
          "prod_node"
        ],
        "title": "ChangeLog",
        "type": "object"
      },
...
      "Edge": {
        "description": "Model representing an edge/relationship in a graph",
        "properties": {
          "created_at": {
            "description": "When the relationship was created",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier for the relationship",
            "title": "Id",
            "type": "string"
          },
          "properties": {
            "description": "Properties of the relationship",
            "title": "Properties",
            "type": "object"
          },
          "source": {
            "description": "ID of the source node",
            "title": "Source",
            "type": "string"
          },
          "target": {
            "description": "ID of the target node",
            "title": "Target",
            "type": "string"
          },
          "type": {
            "description": "Type of the relationship",
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "description": "When the relationship was last updated",
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "target",
          "type"
        ],
        "title": "Edge",
        "type": "object"
      },
...
      "GraphResponse": {
        "description": "Model representing a graph response",
        "properties": {
          "edges": {
            "description": "Relationships in the graph",
            "items": {
              "$ref": "#/components/schemas/Edge"
            },
            "title": "Edges",
            "type": "array"
          },
          "metadata": {
            "description": "Metadata about the graph",
            "title": "Metadata",
            "type": "object"
          },
          "nodes": {
            "description": "Nodes in the graph",
            "items": {
              "$ref": "#/components/schemas/Node"
            },
            "title": "Nodes",
            "type": "array"
          },
          "total_edges": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 0,
            "description": "Total number of relationships for these nodes",
            "title": "Total Edges"
          },
          "total_nodes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 0,
            "description": "Total number of nodes with this label",
            "title": "Total Nodes"
          }
        },
        "title": "GraphResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MergeInitResponse": {
        "description": "Response for merge initialization",
        "properties": {
          "merge_id": {
            "title": "Merge Id",
            "type": "string"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MergeStatus"
          }
        },
        "required": [
          "merge_id",
          "status",
          "start_time"
        ],
        "title": "MergeInitResponse",
        "type": "object"
      },
      "MergeStatus": {
        "description": "Status of a merge operation",
        "enum": [
          "started",
          "auto_resolve",
          "human_review",
          "ready_to_merge",
          "merge_in_progress",
          "completed",
          "failed",
          "cancelled",
          "not_found"
        ],
        "title": "MergeStatus",
        "type": "string"
      },
...
      "Node": {
        "description": "Model representing a node in a graph",
        "properties": {
          "created_at": {
            "description": "When the node was created",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier for the node",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "description": "Primary label for the node",
            "title": "Label",
            "type": "string"
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional labels for the node",
            "title": "Labels"
          },
          "properties": {
            "description": "Properties of the node",
            "title": "Properties",
            "type": "object"
          },
          "type": {
            "description": "Type of the node",
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "description": "When the node was last updated",
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "Node",
        "type": "object"
      },
...
      "ResolutionStrategy": {
        "enum": [
          "KEEP_STAGING",
          "KEEP_PRODUCTION",
          "MERGE_VALUES",
          "KEEP_BOTH"
        ],
        "title": "ResolutionStrategy",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Graphora API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/merge/graph/{merge_id}/{transform_id}": {
      "get": {
        "description": "Retrieve nodes by transform ID and their relationships",
        "operationId": "get_graph_by_merge_id_api_v1_merge_graph__merge_id___transform_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "merge_id",
            "required": true,
            "schema": {
              "title": "Merge Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Graph By Merge Id",
        "tags": [
          "Merge"
        ]
      }
    },
    "/api/v1/merge/statistics/{merge_id}": {
      "get": {
        "description": "Get detailed statistics of a merge operation",
        "operationId": "get_merge_statistics_api_api_v1_merge_statistics__merge_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "merge_id",
            "required": true,
            "schema": {
              "title": "Merge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Merge Statistics Api Api V1 Merge Statistics  Merge Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Merge Statistics Api",
        "tags": [
          "Merge"
        ]
      }
    },
    "/api/v1/merge/{merge_id}/conflicts": {
      "get": {
        "description": "Get conflicts for a merge process",
        "operationId": "get_conflicts_api_v1_merge__merge_id__conflicts_get",
        "parameters": [
          {
            "in": "path",
            "name": "merge_id",
            "required": true,
            "schema": {
              "title": "Merge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChangeLog"
                  },
                  "title": "Response Get Conflicts Api V1 Merge  Merge Id  Conflicts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Conflicts",
        "tags": [
          "Merge"
        ]
      }
    },
    "/api/v1/merge/{merge_id}/conflicts/{conflict_id}/resolve": {
      "post": {
        "description": "Apply a resolution to a specific conflict",
        "operationId": "resolve_conflict_api_v1_merge__merge_id__conflicts__conflict_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "merge_id",
            "required": true,
            "schema": {
              "title": "Merge Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "conflict_id",
            "required": true,
            "schema": {
              "title": "Conflict Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resolution",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ResolutionStrategy"
            }
          },
          {
            "in": "query",
            "name": "learning_comment",
            "required": true,
            "schema": {
              "title": "Learning Comment",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Changed Props",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Resolve Conflict Api V1 Merge  Merge Id  Conflicts  Conflict Id  Resolve Post",
                  "type": "boolean"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Conflict",
        "tags": [
          "Merge"
        ]
      }
    },
    "/api/v1/merge/{merge_id}/status": {
      "get": {
        "description": "Get current status of a merge process for user",
        "operationId": "get_merge_status_api_api_v1_merge__merge_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "merge_id",
            "required": true,
            "schema": {
              "title": "Merge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Merge Status Api",
        "tags": [
          "Merge"
        ]
      }
    },
    "/api/v1/merge/{session_id}/{transform_id}/start": {
      "post": {
        "description": "Start graph merge process",
        "operationId": "start_merge_api_v1_merge__session_id___transform_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "transform_id",
            "required": true,
            "schema": {
              "title": "Transform Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "merge_id",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Merge Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeInitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Merge",
        "tags": [
          "Merge"
        ]
      }
    },
...
  }
}